requirements.txt
telebot=1.4.1
python-binance>=1.0.12
talib>=0.4.0
